requirements.txt
# Real-Time Content Recommendation System Requirements

# Core API Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.6

# Data Processing & ML
pandas==2.3.1
numpy==2.0.2
scikit-learn==1.3.2

# Web Interface
streamlit==1.47.0
plotly==6.2.0

# Database & Storage
happybase==1.2.0

# Streaming & Messaging
kafka-python==2.2.15

# HTTP Requests & Utilities
requests==2.32.4
httpx==0.25.2

# Big Data Processing
pyspark==3.5.0         # For Hadoop integration

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Caching (optional)
redis==5.0.1

# Code Quality (for development)
black==24.10.0
flake8==6.0.0
pytest==7.4.4
